Todas las estadísticas linkeadas!!
Están a medio colocar, pero ya están todos los accesos hechos.
-> Faltarían el arma y la armadura y la descripción en caso de tener.

+++ Proyecto.02/project/src/RPGLogic.h
@@ -142,7 +142,7 @@ namespace rpgLogic {
 			_weaknesses[type] += change;
 		}
 
-		float getWeakness(damageType type) {	// Devolver la resistencia en función de tipo o parametro
+		size_t getWeakness(damageType type) {	// Devolver la resistencia en función de tipo o parametro
 			return _weaknesses[type];
 		}
 	};
